refactor: add commands folder
