+++ src/operstors/basicOperation.mjs
@@ -1,10 +1,6 @@
 import { createReadStream, createWriteStream, promises as fsPromises } from 'fs';
 import { pipeline } from 'stream';
-import { createHash } from 'crypto';
-import { brotliCompress, brotliDecompress } from 'zlib';
-import { EOL, cpus, homedir, userInfo, arch } from 'os';
-import path from 'path';
-import FileManager from './fileManager.mjs';
+import FileManager from '../fileManager.mjs';
 
 export const fileOperator = {
     copyStream : (source, destination) => {
