+++ src/operstors/nwd.mjs
@@ -1,10 +1,6 @@
-import { createReadStream, createWriteStream, promises as fsPromises } from 'fs';
-import { pipeline } from 'stream';
-import { createHash } from 'crypto';
-import { brotliCompress, brotliDecompress } from 'zlib';
-import { EOL, cpus, homedir, userInfo, arch } from 'os';
+import { promises as fsPromises } from 'fs';
 import path from 'path';
-import FileManager from './fileManager.mjs';
+import FileManager from '../fileManager.mjs';
 
 export const navigator = {
     navigateUp : () => {
