+++ src/operstors/osOperator.mjs
@@ -1,0 +1,11 @@
+import { createReadStream, createWriteStream, promises as fsPromises } from 'fs';
+import { pipeline } from 'stream';
+import { createHash } from 'crypto';
+import { brotliCompress, brotliDecompress } from 'zlib';
+import { EOL, cpus, homedir, userInfo, arch } from 'os';
+import path from 'path';
+import FileManager from '../fileManager.mjs';
+
+export const osOperator = {
+    
+}
